Resolving circular import error correctly.
Correctly using StrEnum for comparisons.
